Create a routes.py file for better format and for testing purpose
